requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
astroid==3.3.8
    # via pylint
bcrypt==4.2.1
    # via -r requirements.in
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
coverage==7.6.10
    # via -r requirements.in
dill==0.3.9
    # via pylint
docker==7.1.0
    # via -r requirements.in
fastapi==0.115.8
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via -r requirements.in
humanize==4.11.0
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
isort==6.0.0
    # via pylint
jinja2==3.1.5
    # via -r requirements.in
markupsafe==3.0.2
    # via jinja2
mccabe==0.7.0
    # via pylint
packaging==24.2
    # via pytest
peewee==3.17.8
    # via -r requirements.in
platformdirs==4.3.6
    # via pylint
pluggy==1.5.0
    # via pytest
psutil==7.0.0
    # via -r requirements.in
pydantic==2.10.6
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.7.1
    # via -r requirements.in
pylint==3.3.4
    # via -r requirements.in
pytest==8.3.4
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.25.3
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via -r requirements.in
pyyaml==6.0.2
    # via uvicorn
redis==5.2.1
    # via -r requirements.in
requests==2.32.3
    # via docker
ruff==0.9.9
    # via -r requirements.in
schedule==1.2.2
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
starlette==0.45.3
    # via fastapi
supervisor==4.2.5
    # via -r requirements.in
tomlkit==0.13.2
    # via pylint
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
urllib3==2.3.0
    # via
    #   docker
    #   requests
uvicorn[standard]==0.34.0
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.4
    # via uvicorn
websockets==14.2
    # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
